Update Charm to avoid OpenSSL issue.
The existing commit of `charm` has a [known issue](https://github.com/JHUISI/charm/issues/209)
with OpenSSL 1.1.1 (i.e. the version that supports TLS 1.3). As such, to
use the the fhipe code, one would need to locally downgrade their
version of OpenSSL. However, this is very inconvenient and strongly
discouraged. Instead, updating charm to the dev branch, which supports
the latest version of OpenSSL, fixes the issue "upstream".